+++ optimize/util/optimize-commons/src/test/java/io/camunda/optimize/upgrade/os/OpenSearchClientBuilderTest.java
@@ -9,7 +9,7 @@ package io.camunda.optimize.upgrade.os;
 
 import static org.mockserver.integration.ClientAndServer.startClientAndServer;
 
-import io.camunda.db.se.config.PluginConfiguration;
+import io.camunda.db.search.engine.config.PluginConfiguration;
 import io.camunda.optimize.service.util.configuration.ConfigurationService;
 import io.camunda.optimize.service.util.configuration.OpenSearchConfiguration;
 import io.camunda.optimize.service.util.configuration.db.DatabaseConnection;
@@ -222,12 +222,15 @@ class OpenSearchClientBuilderTest {
     private static final NoopCallback INSTANCE = new NoopCallback();
 
     @Override
-    public void completed(final SimpleHttpResponse result) {}
+    public void completed(final SimpleHttpResponse result) {
+    }
 
     @Override
-    public void failed(final Exception ex) {}
+    public void failed(final Exception ex) {
+    }
 
     @Override
-    public void cancelled() {}
+    public void cancelled() {
+    }
   }
 }
