+++ search/search-client-connect/src/test/java/io/camunda/search/connect/es/ElasticsearchConnectorTest.java
@@ -95,15 +95,12 @@ class ElasticsearchConnectorTest {
     private static final NoopCallback INSTANCE = new NoopCallback();
 
     @Override
-    public void completed(final HttpResponse result) {
-    }
+    public void completed(final HttpResponse result) {}
 
     @Override
-    public void failed(final Exception ex) {
-    }
+    public void failed(final Exception ex) {}
 
     @Override
-    public void cancelled() {
-    }
+    public void cancelled() {}
   }
 }
